Fixed bug with deletion file used as last clusterization

+++ app/narwos.py
@@ -395,6 +395,22 @@ def delete_file():
         logger.debug(f'File {filename} does not exist in File Storage.')
         return redirect(url_for("index", delete_failed_message=f'File {filename} does not exist in File Storage.'))
     
+    files_used_for_clustering = get_rows_cardinalities(
+        path_to_cardinalities_file=PATH_TO_ROWS_CARDINALITIES
+    )
+
+    forbidden_files_to_delete = \
+        list(files_used_for_clustering.get(USED_AS_BASE_KEY, {}).keys()) \
+        + list(files_used_for_clustering.get(ONLY_CLASSIFIED_KEY, {}).keys())
+    
+    if filename in forbidden_files_to_delete:
+        return redirect(url_for("index", delete_failed_message=f"""
+            Deletion of file {filename} is forbidden - it has been used in last clusterization process, 
+            cluster again with other files and then delete this file"""))
+    
+    logger.debug(f'{forbidden_files_to_delete=}')
+    logger.debug(f'{filename=}')
+    
     try:
         os.remove(file_path)
     except Exception as e:
@@ -573,13 +589,16 @@ def get_empty_contents():
 @app.route('/show_clusters', methods=['GET'])
 def show_clusters():
 
+    try:
+        df = read_file(PATH_TO_CURRENT_DF)
+    except FileNotFoundError:
+        return redirect(url_for("index", no_current_df_message=f"Analysis page will be available after first clusterization process - current_df file is missing")) 
+
     message = request.args.get("message")
     update_clusters_new_file_message = request.args.get("update_clusters_new_file_message")
     update_clusters_existing_file_message = request.args.get("update_clusters_existing_file_message")
     update_clusters_existing_file_no_file_message = request.args.get("update_clusters_existing_file_no_file_message")
     update_clusters_new_file_no_file_message = request.args.get("update_clusters_new_file_no_file_message")
-
-    df = read_file(PATH_TO_CURRENT_DF)
 
     if isinstance(df, pd.DataFrame):
